# shared/api-contract.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "静记 API Contract",
    "version": "1.0.0",
    "description": "三端共享的 API 契约文件，用于类型同步和文档生成"
  },
  "paths": {
...
    "/api/statistics/summary": {
      "get": {
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "startDate", "in": "query", "required": true, "schema": { "type": "string", "format": "date" } },
          { "name": "endDate", "in": "query", "required": true, "schema": { "type": "string", "format": "date" } }
        ],
        "responses": {
          "200": {
            "description": "统计摘要",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "totalIncome": { "type": "number" },
                    "totalExpense": { "type": "number" },
                    "balance": { "type": "number" },
                    "previousBalance": { "type": "number" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/statistics/monthly-trend": {
      "get": {
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "months", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 24 } },
          { "name": "endDate", "in": "query", "schema": { "type": "string", "format": "date" } },
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["income", "expense"] } }
        ],
        "responses": {
          "200": {
            "description": "趋势数据",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "month": { "type": "string" },
                      "income": { "type": "number" },
                      "expense": { "type": "number" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/export/excel": {
      "get": {
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["income", "expense"] } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "startDate", "in": "query", "schema": { "type": "string" } },
          { "name": "endDate", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Excel 文件下载",
            "content": {
              "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": {}
            }
          }
        }
      }
    },
    "/api/export/pdf": {
      "get": {
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["income", "expense"] } },
          { "name": "category", "in": "query", "schema": { "type": "string" } },
          { "name": "startDate", "in": "query", "schema": { "type": "string" } },
          { "name": "endDate", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "PDF 文件下载",
            "content": {
              "application/pdf": {}
            }
          }
        }
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT-like-token"
      }
    }
  }
}